Refactor: move functions to utils file
